Question generation workflow & ctrl-flow domain fixes (#310)
* Generator: using generated question files [implementation in progress]

* minor code style corrections (by IDE)

* Minor bugfix and debug print of cmd before run

* Fix Jena manifest issue: build jar with maven-shade-plugin

* Minor fixes on expr generator

* Seems to fix problems with expr generator (run as child java process)

* add .bat file running question generator for expressions

* get rid of commented hardcode

* [fix] Rename domain shortname: control_flow -> ctrl_flow (align with existing questions)

* Don't run JobRunr recurrently (the mode is broken so far)

* [fix] Using QR-log with correct values (filled by domain)

* Dealing with qr-logs while saving new questions

* rename method: runGeneration4Expr -> runQuestionGeneration

* fix background-server's initialize errors

* hide unused dependency in pom.xml

* Clear directories of downloaded repositories (not rm completely)

* configure flags for recurrent job

* Add run mode and job interval to config

---------

+++ modules/core/src/main/java/org/vstu/compprehension/models/entities/QuestionMetadataEntity.java
@@ -44,8 +44,9 @@ public class QuestionMetadataEntity {
     @Column(name = "q_data_graph")
     private String qDataGraph;
 
+    /** workaround due to change in field name in db/json.
+     * Normally, use `setQDataGraph(path)`. */
     public void setQDataGraphPath(String path) {
-        // workaround due to change in field name in db/json
         setQDataGraph(path);
     }
 
@@ -118,7 +119,7 @@ public class QuestionMetadataEntity {
     @Builder.Default
     @Type(type = "json")
     @Column(name = "qrlog_ids", columnDefinition = "json")
-    private List<Integer> qrlogIds = null;
+    private List<Long> qrlogIds = null;
 
     @Column(name = "date_created")
     private Date dateCreated;
